refactor: separate fetcher and app function
