+++ node/pkg/fetcher/fetcher_test.go
@@ -3,8 +3,6 @@ package fetcher
 
 import (
 	"context"
-	"encoding/json"
-	"net/http"
 	"strconv"
 	"testing"
 	"time"
@@ -12,9 +10,6 @@ import (
 	"bisonai.com/orakl/node/pkg/admin/tests"
 	"bisonai.com/orakl/node/pkg/db"
 	"github.com/stretchr/testify/assert"
-
-	"github.com/elazarl/goproxy"
-	"github.com/rs/zerolog/log"
 )
 
 const WAIT_SECONDS = 4 * time.Second
@@ -27,7 +22,7 @@ func TestFetcherInitialize(t *testing.T) {
 	}
 	defer clean()
 
-	app := testItems.fetcher
+	app := testItems.app
 
 	err = app.initialize(ctx)
 	if err != nil {
@@ -49,7 +44,7 @@ func TestFetcherRun(t *testing.T) {
 	}
 	defer clean()
 
-	app := testItems.fetcher
+	app := testItems.app
 
 	err = app.initialize(ctx)
 	if err != nil {
@@ -119,7 +114,7 @@ func TestFetcherFetcherStart(t *testing.T) {
 	}
 	defer clean()
 
-	app := testItems.fetcher
+	app := testItems.app
 
 	err = app.initialize(ctx)
 	if err != nil {
@@ -186,7 +181,7 @@ func TestFetcherFetcherStop(t *testing.T) {
 	}
 	defer clean()
 
-	app := testItems.fetcher
+	app := testItems.app
 
 	err = app.initialize(ctx)
 	if err != nil {
@@ -258,7 +253,7 @@ func TestFetcherFetcherStartById(t *testing.T) {
 	}
 	defer clean()
 
-	app := testItems.fetcher
+	app := testItems.app
 
 	err = app.initialize(ctx)
 	if err != nil {
@@ -268,7 +263,7 @@ func TestFetcherFetcherStartById(t *testing.T) {
 	app.subscribe(ctx)
 
 	for _, fetcher := range app.Fetchers {
-		result, requestErr := tests.PostRequest[Adapter](testItems.app, "/api/v1/adapter/activate/"+strconv.FormatInt(fetcher.ID, 10), nil)
+		result, requestErr := tests.PostRequest[Adapter](testItems.admin, "/api/v1/adapter/activate/"+strconv.FormatInt(fetcher.ID, 10), nil)
 		if requestErr != nil {
 			t.Fatalf("error starting adapter: %v", requestErr)
 		}
@@ -289,7 +284,7 @@ func TestFetcherFetcherStopById(t *testing.T) {
 	}
 	defer clean()
 
-	app := testItems.fetcher
+	app := testItems.app
 
 	err = app.initialize(ctx)
 	if err != nil {
@@ -305,7 +300,7 @@ func TestFetcherFetcherStopById(t *testing.T) {
 	}
 
 	for _, fetcher := range app.Fetchers {
-		result, requestErr := tests.PostRequest[Adapter](testItems.app, "/api/v1/adapter/deactivate/"+strconv.FormatInt(fetcher.ID, 10), nil)
+		result, requestErr := tests.PostRequest[Adapter](testItems.admin, "/api/v1/adapter/deactivate/"+strconv.FormatInt(fetcher.ID, 10), nil)
 		if requestErr != nil {
 			t.Fatalf("error stopping adapter: %v", requestErr)
 		}
@@ -315,132 +310,5 @@ func TestFetcherFetcherStopById(t *testing.T) {
 
 	for _, fetcher := range app.Fetchers {
 		assert.False(t, fetcher.isRunning)
-	}
-}
-
-func TestFetcherFetch(t *testing.T) {
-	ctx := context.Background()
-	clean, testItems, err := setup(ctx)
-	if err != nil {
-		t.Fatalf("error setting up test: %v", err)
-	}
-	defer clean()
-
-	app := testItems.fetcher
-
-	err = app.initialize(ctx)
-	if err != nil {
-		t.Fatalf("error initializing fetcher: %v", err)
-	}
-
-	for _, fetcher := range app.Fetchers {
-		result, err := app.fetch(*fetcher)
-		if err != nil {
-			t.Fatalf("error fetching: %v", err)
-		}
-		assert.Greater(t, len(result), 0)
-	}
-}
-
-func TestFetcherFetchProxy(t *testing.T) {
-	ctx := context.Background()
-	clean, testItems, err := setup(ctx)
-	if err != nil {
-		t.Fatalf("error setting up test: %v", err)
-	}
-	defer clean()
-
-	app := testItems.fetcher
-
-	proxyServer := goproxy.NewProxyHttpServer()
-	srv := &http.Server{
-		Addr:    ":8088",
-		Handler: proxyServer,
-	}
-	go func() {
-		if proxyServeErr := srv.ListenAndServe(); proxyServeErr != http.ErrServerClosed {
-			// Unexpected server shutdown
-			log.Fatal().Err(proxyServeErr).Msg("unexpected server shutdown")
-		}
-	}()
-
-	proxy, err := tests.PostRequest[Proxy](testItems.app, "/api/v1/proxy", map[string]any{"protocol": "http", "host": "localhost", "port": 8088})
-	if err != nil {
-		t.Fatalf("error creating proxy: %v", err)
-	}
-
-	err = app.initialize(ctx)
-	if err != nil {
-		t.Fatalf("error initializing fetcher: %v", err)
-	}
-
-	for _, fetcher := range app.Fetchers {
-		result, fetchErr := app.fetch(*fetcher)
-		if fetchErr != nil {
-			t.Fatalf("error fetching: %v", fetchErr)
-		}
-		assert.Greater(t, len(result), 0)
-	}
-
-	_, err = tests.DeleteRequest[Proxy](testItems.app, "/api/v1/proxy/"+strconv.FormatInt(proxy.ID, 10), nil)
-	if err != nil {
-		t.Fatalf("error cleaning up proxy: %v", err)
-	}
-
-	if err = srv.Shutdown(ctx); err != nil {
-		log.Fatal().Err(err).Msg("unexpected server shutdown")
-	}
-
-}
-
-func TestFetcherFetchAndInsertAdapter(t *testing.T) {
-	ctx := context.Background()
-	clean, testItems, err := setup(ctx)
-	if err != nil {
-		t.Fatalf("error setting up test: %v", err)
-	}
-	defer clean()
-
-	app := testItems.fetcher
-
-	app.initialize(ctx)
-
-	for _, fetcher := range app.Fetchers {
-		app.fetchAndInsert(ctx, *fetcher)
-	}
-	if err != nil {
-		t.Fatalf("error running adapter: %v", err)
-	}
-
-	for _, fetcher := range app.Fetchers {
-		pgResult, err := db.QueryRow[Aggregate](ctx, "SELECT * FROM local_aggregates WHERE name = @name", map[string]any{"name": fetcher.Name})
-		if err != nil {
-			t.Fatalf("error reading from db: %v", err)
-		}
-		assert.NotNil(t, pgResult)
-
-		feedPgResult, err := db.QueryRows[FeedDataFromDB](ctx, "SELECT * FROM feed_data WHERE adapter_id = @adapter_id", map[string]any{"adapter_id": fetcher.Adapter.ID})
-		if err != nil {
-			t.Fatalf("error reading from db: %v", err)
-		}
-		assert.Greater(t, len(feedPgResult), 0)
-
-		rdbResult, err := db.Get(ctx, "localAggregate:"+fetcher.Name)
-		if err != nil {
-			t.Fatalf("error reading from redis: %v", err)
-		}
-		assert.NotNil(t, rdbResult)
-		var redisAgg redisAggregate
-		err = json.Unmarshal([]byte(rdbResult), &redisAgg)
-		if err != nil {
-			t.Fatalf("error unmarshalling from redis: %v", err)
-		}
-		assert.NotNil(t, redisAgg)
-		assert.NotNil(t, redisAgg.Value)
-
-		err = db.Del(ctx, "localAggregate:"+fetcher.Name)
-		if err != nil {
-			t.Fatalf("error removing from redis: %v", err)
-		}
 	}
 }
